requirements.txt
aiogram
APScheduler
python-decouple
asyncpg